# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "VSB 服务端 API 文档",
        "title": "VSB Server API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/product/category/create": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "创建新商品分类，需要 Bearer Token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "商品分类"
                ],
                "summary": "创建商品分类",
                "parameters": [
                    {
                        "description": "分类信息",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/category.CategoryRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "创建成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/category.CategoryResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "参数错误或分类已存在",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/product/category/delete/{id}": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "根据分类 ID 删除商品分类，需要 Bearer Token",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "商品分类"
                ],
                "summary": "删除商品分类",
                "parameters": [
                    {
                        "type": "string",
                        "description": "分类 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "删除成功",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/product/category/list": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "获取全部商品分类，需要 Bearer Token",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "商品分类"
                ],
                "summary": "获取商品分类列表",
                "responses": {
                    "200": {
                        "description": "分类列表",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/definitions/response.PageData"
                                                },
                                                {
                                                    "type": "object",
                                                    "properties": {
                                                        "list": {
                                                            "type": "array",
                                                            "items": {
                                                                "$ref": "#/definitions/category.CategoryResponse"
                                                            }
                                                        }
                                                    }
                                                }
                                            ]
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/product/category/update/{id}": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "根据分类 ID 更新商品分类，空字段表示不修改，需要 Bearer Token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "商品分类"
                ],
                "summary": "更新商品分类",
                "parameters": [
                    {
                        "type": "string",
                        "description": "分类 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "更新信息",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/category.CategoryUpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "更新成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/category.CategoryResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "category.CategoryRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "description": {
                    "description": "分类描述（多语言）",
                    "allOf": [
                        {
                            "$ref": "#/definitions/i18n.Locale"
                        }
                    ]
                },
                "name": {
                    "description": "分类名称（多语言）",
                    "allOf": [
                        {
                            "$ref": "#/definitions/i18n.Locale"
                        }
                    ]
                },
                "parentId": {
                    "description": "父分类 ID，为空表示顶级分类",
                    "type": "string"
                }
            }
        },
        "category.CategoryResponse": {
            "type": "object",
            "properties": {
                "categoryId": {
                    "description": "分类 ID",
                    "type": "string"
                },
                "createdAt": {
                    "description": "创建时间",
                    "type": "string"
                },
                "description": {
                    "description": "分类描述（多语言）",
                    "allOf": [
                        {
                            "$ref": "#/definitions/i18n.Locale"
                        }
                    ]
                },
                "name": {
                    "description": "分类名称（多语言）",
                    "allOf": [
                        {
                            "$ref": "#/definitions/i18n.Locale"
                        }
                    ]
                },
                "parentId": {
                    "description": "父分类 ID，为空表示顶级分类",
                    "type": "string"
                },
                "status": {
                    "description": "分类状态",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.StatusEnum"
                        }
                    ]
                },
                "updatedAt": {
                    "description": "更新时间",
                    "type": "string"
                }
            }
        },
        "category.CategoryUpdateRequest": {
            "type": "object",
            "properties": {
                "categoryId": {
                    "description": "分类 ID，以路径参数为准",
                    "type": "string"
                },
                "description": {
                    "description": "分类描述（多语言）",
                    "allOf": [
                        {
                            "$ref": "#/definitions/i18n.Locale"
                        }
                    ]
                },
                "name": {
                    "description": "分类名称（多语言）",
                    "allOf": [
                        {
                            "$ref": "#/definitions/i18n.Locale"
                        }
                    ]
                },
                "parentId": {
                    "description": "父分类 ID",
                    "type": "string"
                },
                "status": {
                    "description": "分类状态",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.StatusEnum"
                        }
                    ]
                }
            }
        },
...
        "i18n.Locale": {
            "type": "object",
            "additionalProperties": {
                "type": "string"
            }
        },
...
        "models.StatusEnum": {
            "type": "string",
            "enum": [
                "active",
                "unlisted",
                "listed",
                "inactive",
                "deleted",
                "expired",
                "pending",
                "approved",
                "disabled",
                "enabled"
            ],
            "x-enum-comments": {
                "StatusActive": "活跃中",
                "StatusApproved": "审核通过",
                "StatusDeleted": "已删除",
                "StatusDisabled": "禁用",
                "StatusEnabled": "启用",
                "StatusExpired": "已过期",
                "StatusInactive": "已下架",
                "StatusListed": "已上架",
                "StatusPending": "待审核",
                "StatusUnlisted": "未上架"
            },
            "x-enum-descriptions": [
                "活跃中",
                "未上架",
                "已上架",
                "已下架",
                "已删除",
                "已过期",
                "待审核",
                "审核通过",
                "禁用",
                "启用"
            ],
            "x-enum-varnames": [
                "StatusActive",
                "StatusUnlisted",
                "StatusListed",
                "StatusInactive",
                "StatusDeleted",
                "StatusExpired",
                "StatusPending",
                "StatusApproved",
                "StatusDisabled",
                "StatusEnabled"
            ]
        },
        "response.PageData": {
            "type": "object",
            "properties": {
                "list": {
                    "description": "当前页数据列表"
                },
                "pageIndex": {
                    "description": "当前页码，从 1 开始",
                    "type": "integer"
                },
                "pageSize": {
                    "description": "每页条数",
                    "type": "integer"
                },
                "total": {
                    "description": "总条数",
                    "type": "integer"
                }
            }
        },
        "response.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "状态码",
                    "type": "integer"
                },
                "data": {
                    "description": "数据"
                },
                "message": {
                    "description": "消息",
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Bearer JWT token，格式：Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}